pyproject.toml
[project]
name = "nonebot_plugin_course_schedule"
version = "0.1.0"
description = "课程表插件,可以视奸群友在上什么课"
authors = [
    {name = "Koileo", email = "koileo@outlook.com"},
]
dependencies = [
  "nonebot2>=2.0.0",
  "nonebot-adapter-onebot>=2.0.0",
  "aiohttp>=3.8.0",
  "icalendar>=5.0.0",
  "python-dateutil>=2.8.0",
  "Pillow>=9.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["nonebot_plugin_course_schedule"]

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.4.2",
    "isort>=5.13.2",
    "ruff>=0.4.6",
]

[tool.pdm.scripts]
format = { composite = ["isort . ","black . ","ruff check ."] }
test = "python -c 'print(\">>> Just \\\"pytest -W ignore -s\\\" when you complete your testsuite\")'"


[tool.black]
line-length = 90
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 90
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "T201", "E731", "E402"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"
